+++ src/index.ts
@@ -4,7 +4,7 @@ import type { Linter } from "eslint";
 import type { Options } from "./types.js";
 
 import { defaultOptions } from "./defaults.js";
-import { environment, stylistic, base, node, types, react, importPlugin, tests } from "./rules/index.js";
+import { environment, stylistic, base, node, types, react, importPlugin, tests, todoTaskId } from "./rules/index.js";
 
 /**
  * Gets the eslint configuration based on the options.
@@ -21,6 +21,7 @@ const getEslintConfig = (options: Options = {}): Linter.FlatConfig[] => {
         ...types.get(mergedOptions),
         ...react.get(mergedOptions),
         ...importPlugin.get(mergedOptions),
+        ...todoTaskId.get(mergedOptions),
         tests.get(mergedOptions),
     ].filter(truthy);
 };