+++ src/rules/index.ts
@@ -5,4 +5,5 @@ export * as node from "./node.js";
 export * as react from "./react.js";
 export * as stylistic from "./stylistic.js";
 export * as tests from "./tests.js";
+export * as todoTaskId from "./todoTaskId.js";
 export * as types from "./types.js";