+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [UNRELEASED]
 (nothing yet)
 
+## [0.2.0] - 2024-12-28
+### Added
+- `@ezez/eslint-plugin-todo-task-id` plugin (disabled by default)
+### Dev
+- bump import plugin to fix some issues
+
 ## [0.1.0] - 2024-11-07
 ### Changed
 - `eslint` is no longer peer dependency